Plant_annotation/requirements.txt
openai>=1.40.0
streamlit>=1.36.0
pillow>=10.3.0
pandas>=2.2.2
python-dotenv>=1.0.1
